requirements.txt
configparser==4.0.2
mysql-connector==2.2.9
pandas==0.25.3
requests==2.22.0
